address review comments:
- move Snapshot classes into the corresponding Model
- include SkaterModels in snapshots
