+++ src/com/carolinarollergirls/scoreboard/Team.java
@@ -34,7 +34,6 @@ public interface Team extends ScoreBoardEventProvider
 	public int getTimeouts();
 	public int getOfficialReviews();
 
-	public boolean inJam();
 	public boolean inTimeout();
 	public boolean inOfficialReview();
 	public boolean retainedOfficialReview();
