+++ src/com/carolinarollergirls/scoreboard/model/TeamModel.java
@@ -8,12 +8,13 @@ package com.carolinarollergirls.scoreboard.model;
  * See the file COPYING for details.
  */
 
+import java.util.HashMap;
 import java.util.List;
 
 import com.carolinarollergirls.scoreboard.PositionNotFoundException;
 import com.carolinarollergirls.scoreboard.SkaterNotFoundException;
 import com.carolinarollergirls.scoreboard.Team;
-import com.carolinarollergirls.scoreboard.snapshots.TeamSnapshot;
+import com.carolinarollergirls.scoreboard.model.SkaterModel.SkaterSnapshot;
 
 public interface TeamModel extends Team
 {
@@ -59,7 +60,6 @@ public interface TeamModel extends Team
 	public void changeOfficialReviews(int reviews);
 	public void resetTimeouts(boolean gameStart);
 
-	public void setInJam(boolean inJam);
 	public void setInTimeout(boolean in_timeouts);
 	public void setInOfficialReview(boolean in_official_review);
 	public void setRetainedOfficialReview(boolean retained_official_review);
@@ -79,7 +79,7 @@ public interface TeamModel extends Team
 
 	public void setStarPass(boolean starPass);
 
-  public void penalty(String skaterId, String penaltyId, boolean fo_exp, int period, int jam, String code);
+	public void penalty(String skaterId, String penaltyId, boolean fo_exp, int period, int jam, String code);
 
 	public static interface AlternateNameModel extends AlternateName {
 		public void setName(String n);
@@ -91,5 +91,46 @@ public interface TeamModel extends Team
 		public void setColor(String c);
 
 		public TeamModel getTeamModel();
+	}
+	
+	public class TeamSnapshot {
+		public TeamSnapshot(TeamModel team) {
+			id = team.getId();
+			score = team.getScore();
+			lastscore = team.getLastScore();
+			timeouts = team.getTimeouts();
+			officialReviews = team.getOfficialReviews();
+			leadJammer = team.getLeadJammer();
+			starPass = team.isStarPass();
+			in_timeout = team.inTimeout();
+			in_official_review = team.inOfficialReview();
+			skaterSnapshots = new HashMap<String, SkaterSnapshot>();
+			for (SkaterModel skater : team.getSkaterModels()) {
+				skaterSnapshots.put(skater.getId(), new SkaterSnapshot(skater));
+			}
+		}
+
+		public String getId() { return id; }
+		public int getScore() { return score;}
+		public int getLastScore() { return lastscore; }
+		public int getTimeouts() { return timeouts; }
+		public int getOfficialReviews() { return officialReviews; }
+		public String getLeadJammer() { return leadJammer; }
+		public boolean getStarPass() { return starPass; }
+		public boolean inTimeout() { return in_timeout; }
+		public boolean inOfficialReview() { return in_official_review; }
+		public HashMap<String, SkaterSnapshot> getSkaterSnapshots() { return skaterSnapshots; }
+		public SkaterSnapshot getSkaterSnapshot(String skater) { return skaterSnapshots.get(skater); }
+		
+		protected String id;
+		protected int score;
+		protected int lastscore;
+		protected int timeouts;
+		protected int officialReviews;
+		protected String leadJammer;
+		protected boolean starPass;
+		protected boolean in_timeout;
+		protected boolean in_official_review;
+		protected HashMap<String, SkaterSnapshot> skaterSnapshots;
 	}
 }
